thymeleaf page 띄우기 성공...

+++ src/main/java/com/solution/appsolute/dto/approval/Approval.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @Builder
 public class Approval {
 
-    private int approvalNum;
+    private long approvalNum;
     private int appCheck;
     private LocalDateTime approvalDate;
     private int appContent;
